requirements.txt
colorama==0.4.4
discord.py==2.0.1
psutil==5.9.2
python-dotenv==0.21.0
requests==2.25.1
selenium==4.5.0
